fix(stylex_transform): widen the code-frame span cache to 128 bits too
The count of consumers that act on a bare hash hit was wrong. It is four, not
two, and the two that were missed are the ones that fail *visibly*.

The code-frame `span_cache` is read twice on a bare hit -- once keyed by
`compute_cache_key`, once by `compute_key_span_cache_key` -- and what comes back
becomes the `file:line` on `$$css`. A collision there annotates a style with
another style's line number, in the output, where the evaluator's memo at least
fails silently. Both keys were still 64-bit `DefaultHasher`. Widened, which is
why `stable_hash_wide` is public now rather than private to `hash.rs`. They key a
positional hash rather than the structural one -- the point is caching where
something was written, so two identical expressions at different positions must
not share an entry -- so they needed their own derivation.
`compute_key_span_cache_key` is now one tuple hash instead of a dozen sequential
`hash` calls, so a field added to the function cannot be left out of the digest.

Four other things from the same review:

`Hasher::finish` on the wide hasher is not the low half of `finish_wide`. xxh3's
64-bit and 128-bit digests are separate constructions over one stream, so the two
are unrelated numbers; the doc comment claimed otherwise. Corrected and pinned,
because a caller reaching for the familiar `finish` gets a narrower key *and* a
different one.

`Xxh3Default` replaces `Xxh3`, which copies a 192-byte custom secret into every
instance. One instance is built per key, so that copy was the largest fixed cost
of a small key -- another 2-5% on the fold.

`register_source_once` returns `Result` instead of `bool`, restoring the io cause
that `load_code_frame_from_cache_for_state` used to report and now reports again.

`with_memoized_module` replaces the `bool` that said nothing and the four copies
of the same `return Err`. It hands the memoized module to a closure, which is what
lets the borrow end before the resolved span is written back -- the reason the
previous shape returned an owned clone in the first place.

The dependency is BSL-1.0, not BSD-2-Clause as the ADR said. Both are on the
allow-list; the ADR was simply wrong.

+++ crates/stylex-transform/src/shared/structures/state_manager.rs
@@ -283,15 +283,15 @@ impl CallExpressionState {
 #[derive(Clone, Debug, Default)]
 pub(crate) struct CacheState {
   css_property_seen: FxHashMap<String, String>,
-  span_cache: FxHashMap<u64, Span>,
+  span_cache: FxHashMap<u128, Span>,
 }
 
 impl CacheState {
-  fn cached_span(&self, cache_key: u64) -> Option<Span> {
+  fn cached_span(&self, cache_key: u128) -> Option<Span> {
     self.span_cache.get(&cache_key).copied()
   }
 
-  fn insert_span(&mut self, cache_key: u64, span: Span) {
+  fn insert_span(&mut self, cache_key: u128, span: Span) {
     self.span_cache.insert(cache_key, span);
   }
 }
@@ -595,11 +595,11 @@ impl StateManager {
     self.call_expressions.is_member_callee(member)
   }
 
-  pub(crate) fn cached_span(&self, cache_key: u64) -> Option<Span> {
+  pub(crate) fn cached_span(&self, cache_key: u128) -> Option<Span> {
     self.cache.cached_span(cache_key)
   }
 
-  pub(crate) fn insert_cached_span(&mut self, cache_key: u64, span: Span) {
+  pub(crate) fn insert_cached_span(&mut self, cache_key: u128, span: Span) {
     self.cache.insert_span(cache_key, span);
   }
